Implement playGame and playRound
Finish the project by implementing 2 functions: playGame and playRound.
playRound gets in 2 arguments, which is the decision of each player,
and process who wins the round.
playGame plays the entire game, which is 5 rounds.
There could be a minor improvement: right now, after 5 games, 2 players
can draw. That is because we ca make the same decision and make a round
draw. But I will leave it there and call it for a day.